Add line text as marker text and remove line numbers
In order to make ex. class XXX and def xxx working I added the texts as marker texts, and removed the line numbers to get more space for the text. I also shortened the text so it won't get folded in the default Sidebar width. 

I changed the height of the list to make room for some icons next to it and added a box around the current line.

I hope that's okay, if not I'm fine with that too.